refactor: extract component BottomButtons from all variants of ElementNode
