+++ packages/diagram/src/overlays/relationships-of/xyflow/ElementNode.tsx
@@ -1,5 +1,4 @@
-import { ActionIcon, Box, Group, Text as MantineText } from '@mantine/core'
-import { IconFileSymlink, IconTransform, IconZoomScan } from '@tabler/icons-react'
+import { Box, Text as MantineText } from '@mantine/core'
 import { Handle, type NodeProps, Position } from '@xyflow/react'
 import clsx from 'clsx'
 import { deepEqual } from 'fast-equals'
@@ -7,18 +6,10 @@ import { m } from 'framer-motion'
 import { memo } from 'react'
 import { type DiagramState, useDiagramState } from '../../../hooks'
 import { ElementShapeSvg } from '../../../xyflow/nodes/element/ElementShapeSvg'
-import { stopPropagation } from '../../../xyflow/utils'
 import { useOverlayDialog } from '../../OverlayContext'
 import type { XYFlowTypes } from '../_types'
 import * as css from './styles.css'
-
-const Action = ActionIcon.withProps({
-  className: 'nodrag nopan ' + css.navigateBtn,
-  radius: 'md',
-  role: 'button',
-  onDoubleClick: stopPropagation,
-  onPointerDownCapture: stopPropagation
-})
+import { BottomButtons } from '../../../xyflow/nodes/element/BottomButtons'
 
 const Text = MantineText.withProps({
   component: 'div'
@@ -134,39 +125,25 @@ export const ElementNode = memo<ElementNodeProps>(({
             <Text className={css.elementNodeDescription} lineClamp={4}>{element.description}</Text>
           )}
         </Box>
-        <Group className={css.navigateBtnBox}>
-          {navigateTo && onNavigateTo && navigateTo !== currentViewId && (
-            <Action
-              onClick={(event) => {
-                event.stopPropagation()
-                overlay.close(() => onNavigateTo(navigateTo))
-              }}>
-              <IconZoomScan stroke={1.8} style={{ width: '75%' }} />
-            </Action>
-          )}
-          {data.column !== 'subjects' && (
-            <Action
-              onClick={(event) => {
-                event.stopPropagation()
-                overlay.openOverlay({
-                  relationshipsOf: data.fqn
-                })
-              }}>
-              <IconTransform stroke={1.8} style={{ width: '72%' }} />
-            </Action>
-          )}
-          {onOpenSource && (
-            <Action
-              onClick={(event) => {
-                event.stopPropagation()
-                onOpenSource?.({
-                  element: data.fqn
-                })
-              }}>
-              <IconFileSymlink stroke={1.8} style={{ width: '72%' }} />
-            </Action>
-          )}
-        </Group>
+        {/* Navigation */}
+        <BottomButtons
+          onNavigateTo={navigateTo && onNavigateTo && navigateTo !== currentViewId && ((event) => {
+            event.stopPropagation()
+            overlay.close(() => onNavigateTo(navigateTo))
+          })}
+          onOpenRelationships={data.column !== 'subjects' && ((event) => {
+            event.stopPropagation()
+            overlay.openOverlay({
+              relationshipsOf: data.fqn
+            })
+          })}
+          onOpenSource={onOpenSource && ((event) => {
+            event.stopPropagation()
+            onOpenSource?.({
+              element: data.fqn
+            })
+          })}
+        />
       </m.div>
       {ports.left.map(({ id, type }, i) => (
         <Handle
